# internal/docs/swagger.yaml
definitions:
...
  entity.CreatePostRequest:
    properties:
      content:
        additionalProperties:
          type: string
        type: object
      short_content:
        additionalProperties:
          type: string
        type: object
      slug:
        type: string
      status:
        type: boolean
      title:
        additionalProperties:
          type: string
        type: object
      user_id:
        type: integer
    type: object
  entity.CreatePostResponse:
    properties:
      content:
        additionalProperties:
          type: string
        type: object
      files:
        items:
          type: string
        type: array
      id:
        type: integer
      short_content:
        additionalProperties:
          type: string
        type: object
      slug:
        type: string
      status:
        type: boolean
      title:
        additionalProperties:
          type: string
        type: object
      user_id:
        type: integer
    type: object
...
  entity.DeletePostResponse:
    properties:
      message:
        type: string
    type: object
...
  entity.GetPostResponse:
    properties:
      content:
        additionalProperties:
          type: string
        type: object
      files:
        items:
          type: string
        type: array
      id:
        type: integer
      short_content:
        additionalProperties:
          type: string
        type: object
      slug:
        type: string
      status:
        type: boolean
      title:
        additionalProperties:
          type: string
        type: object
      user_id:
        type: integer
    type: object
...
  entity.ListPostResponse:
    properties:
      menus:
        items:
          $ref: '#/definitions/entity.GetPostResponse'
        type: array
      total:
        type: integer
    type: object
...
  entity.UpdatePostColumnsRequest:
    properties:
      fields:
        additionalProperties:
          type: string
        type: object
      id:
        type: integer
    type: object
  entity.UpdatePostRequest:
    properties:
      content:
        additionalProperties:
          type: string
        type: object
      files:
        items:
          type: string
        type: array
      id:
        type: integer
      short_content:
        additionalProperties:
          type: string
        type: object
      slug:
        type: string
      status:
        type: boolean
      title:
        additionalProperties:
          type: string
        type: object
      user_id:
        type: integer
    type: object
  entity.UpdatePostResponse:
    properties:
      content:
        additionalProperties:
          type: string
        type: object
      files:
        items:
          type: string
        type: array
      id:
        type: integer
      short_content:
        additionalProperties:
          type: string
        type: object
      slug:
        type: string
      status:
        type: boolean
      title:
        additionalProperties:
          type: string
        type: object
      user_id:
        type: integer
    type: object
...
  errors.Error:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/post:
    patch:
      consumes:
      - application/json
      description: This API for updating a post columns
      parameters:
      - description: Update Post Columns Model
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/entity.UpdatePostColumnsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UpdatePostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Update Post Columns
      tags:
      - post
    post:
      consumes:
      - application/json
      description: This API for creating a new post
      parameters:
      - description: Create Post Model
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/entity.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.CreatePostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Create Post
      tags:
      - post
    put:
      consumes:
      - application/json
      description: This API for updating a post
      parameters:
      - description: Update Post Model
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/entity.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UpdatePostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Update Post
      tags:
      - post
  /v1/post/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting post with id
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.DeletePostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Delete Post
      tags:
      - post
    get:
      consumes:
      - application/json
      description: This API for getting list with id
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetPostResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Get Post
      tags:
      - post
  /v1/post/list:
    get:
      consumes:
      - application/json
      description: This API for getting list post
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ListPostResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      summary: List Post
      tags:
      - post
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"